requirements.txt
openai
beautifulsoup4
requests
flask
python-dotenv